+++ Api_CreartNino/Models/Compra.cs
@@ -13,7 +13,7 @@ public partial class Compra
 
     public DateOnly? FechaCompra { get; set; }
 
-    public decimal? Total { get; set; }
+    public int? Total { get; set; }
 
    
     public int? IdEstado { get; set; }
