+++ Api_CreartNino/Models/DetallesCompra.cs
@@ -13,9 +13,9 @@ public partial class DetallesCompra
 
     public int? Cantidad { get; set; }
 
-    public decimal? PrecioUnitario { get; set; }
+    public int? PrecioUnitario { get; set; }
 
-    public decimal? Subtotal { get; set; }
+    public int? Subtotal { get; set; }
 
     public virtual Compra? IdCompraNavigation { get; set; }
