+++ Api_CreartNino/Models/Producto.cs
@@ -17,7 +17,7 @@ public partial class Producto
 
     public string? Marca { get; set; }
 
-    public decimal? Precio { get; set; }
+    public int? Precio { get; set; }
 
     public bool? Estado { get; set; }
